+++ migrate/migration.go
@@ -3,9 +3,14 @@ package migrate
 import (
 	"errors"
 	"time"
+
+	"github.com/yaitoo/sqle/shardid"
 )
 
-var ErrInvalidScriptName = errors.New("migrate: invalid script name")
+var (
+	ErrInvalidScriptName  = errors.New("migrate: invalid script name")
+	ErrInvalidRotateRange = errors.New("migrate: invalid rotate range")
+)
 
 type Semver struct {
 	Name       string
@@ -24,10 +29,15 @@ func (s *Semver) Less(i, j int) bool {
 }
 
 type Migration struct {
-	Name          string
-	Rank          int
-	Checksum      string
-	Scripts       string
+	Name     string
+	Rank     int
+	Checksum string
+	Scripts  string
+
+	Rotate      shardid.TableRotate
+	RotateBegin time.Time
+	RotateEnd   time.Time
+
 	MigratedOn    *time.Time
 	ExecutionTime time.Duration
 }