+++ CHANGELOG.md
@@ -5,16 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.1]
+## [1.2.2]
 ### Added
 - added detail logs on migration (#17)
-  
+- added rotated table migration support (#17)
+
+### Fixed
+- fixed rollback issue on `Transaction` (#17)
+
+## [1.2.1] - 2024-02-28
 ### Fixed
 - fixed missed input variable issue (#15)
 - fixed sql.Scanner/driver.Valuer support in BitBool/shardid.ID (#16)
   
 
-## [1.2.0] - 2024-02-36
+## [1.2.0] - 2024-02-26
 ### Added
 - added `BitBool` for mysql bit type (#11)
 - added `sharding` feature (#12)